# docs/openapi.yaml
openapi: 3.0.0
info:
  title: SaaS Boilerplate API
  description: Multi-tenant SaaS platform API for managing tenants, organizations, users, and roles
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@yourdomain.com
    url: https://yourdomain.com/support
  license:
    name: MIT
    url: https://github.com/your-org/saas-boilerplate/blob/main/LICENSE
servers:
  - url: https://api.yourdomain.com/v1
    description: Production server
  - url: https://staging-api.yourdomain.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  schemas:
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: 'Demo Corporation'
        slug:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-z0-9-]+$'
          example: 'demo-corp'
        domain:
          type: string
          format: hostname
          nullable: true
          example: 'demo-corp.com'
        status:
          type: string
          enum: [active, inactive, suspended]
          example: 'active'
        settings:
          type: object
          additionalProperties: true
          example:
            theme: 'light'
            timezone: 'UTC'
        metadata:
          type: object
          additionalProperties: true
          example:
            source: 'api'
            environment: 'production'
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
      required:
        - id
        - name
        - slug
        - status
        - createdAt
        - updatedAt
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440001'
        tenantId:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: 'Engineering Department'
        description:
          type: string
          nullable: true
          example: 'Main engineering team'
        website:
          type: string
          format: uri
          nullable: true
          example: 'https://demo-corp.com/engineering'
        industry:
          type: string
          nullable: true
          example: 'Technology'
        size:
          type: string
          enum: ['1-10', '11-50', '51-200', '201-500', '500+']
          nullable: true
          example: '51-200'
        settings:
          type: object
          additionalProperties: true
          example:
            department: 'engineering'
            budget: 500000
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
      required:
        - id
        - tenantId
        - name
        - createdAt
        - updatedAt
...
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440003'
        tenantId:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: 'Manager'
        description:
          type: string
          nullable: true
          example: 'Team management access'
        permissions:
          type: array
          items:
            type: string
          example: ['users.read', 'users.write', 'organizations.read']
        isSystem:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
      required:
        - id
        - tenantId
        - name
        - permissions
        - createdAt
        - updatedAt
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        total:
          type: integer
          minimum: 0
          example: 150
        page:
          type: integer
          minimum: 1
          example: 2
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: true
      required:
        - data
        - total
        - page
        - limit
        - hasNext
        - hasPrev
    CreateTenantRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: 'New Tenant'
        slug:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-z0-9-]+$'
          example: 'new-tenant'
        settings:
          type: object
          additionalProperties: true
          example:
            theme: 'dark'
            timezone: 'UTC'
        metadata:
          type: object
          additionalProperties: true
          example:
            source: 'api'
      required:
        - name
        - slug
    CreateOrganizationRequest:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: 'Engineering Department'
        description:
          type: string
          nullable: true
          example: 'Main engineering team'
        website:
          type: string
          format: uri
          nullable: true
          example: 'https://example.com/engineering'
        industry:
          type: string
          nullable: true
          example: 'Technology'
        size:
          type: string
          enum: ['1-10', '11-50', '51-200', '201-500', '500+']
          nullable: true
          example: '51-200'
        settings:
          type: object
          additionalProperties: true
          example:
            department: 'engineering'
            budget: 500000
      required:
        - tenantId
        - name
...
    CreateRoleRequest:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: 'Manager'
        description:
          type: string
          nullable: true
          example: 'Team management access'
        permissions:
          type: array
          items:
            type: string
          example: ['users.read', 'users.write', 'organizations.read']
        isSystem:
          type: boolean
          example: false
      required:
        - tenantId
        - name
        - permissions
  parameters:
    TenantIdParam:
      name: tenantId
      in: query
      description: Filter results by tenant ID
      required: false
      schema:
        type: string
        format: uuid
        example: '550e8400-e29b-41d4-a716-446655440000'
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    SearchParam:
      name: search
      in: query
      description: Search term for filtering results
      required: false
      schema:
        type: string
        example: 'demo'
    StatusParam:
      name: status
      in: query
      description: Filter by status
      required: false
      schema:
        type: string
        enum: [active, inactive, suspended]
        example: 'active'
    IdPathParam:
      name: id
      in: path
      description: Resource ID
      required: true
      schema:
        type: string
        format: uuid
        example: '550e8400-e29b-41d4-a716-446655440000'
paths:
  /tenants:
    get:
      summary: List all tenants
      description: Retrieve a paginated list of tenants with optional filtering
      tags:
        - Tenants
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: List of tenants retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new tenant
      description: Create a new tenant with the provided configuration
      tags:
        - Tenants
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tenants/{id}:
    get:
      summary: Get tenant by ID
      description: Retrieve a specific tenant by its ID
      tags:
        - Tenants
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Tenant retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update tenant
      description: Update an existing tenant with new data
      tags:
        - Tenants
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CreateTenantRequest'
                - type: object
                  description: Partial update of tenant data
      responses:
        '200':
          description: Tenant updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete tenant
      description: Deactivate a tenant (soft delete)
      tags:
        - Tenants
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Tenant deactivated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /organizations:
    get:
      summary: List organizations
      description: Retrieve a paginated list of organizations for a specific tenant
      tags:
        - Organizations
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: List of organizations retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create organization
      description: Create a new organization within a tenant
      tags:
        - Organizations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /organizations/{id}:
    get:
      summary: Get organization by ID
      description: Retrieve a specific organization by its ID
      tags:
        - Organizations
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
        - $ref: '#/components/parameters/TenantIdParam'
      responses:
        '200':
          description: Organization retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /roles:
    get:
      summary: List roles
      description: Retrieve a paginated list of roles for a specific tenant
      tags:
        - Roles
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: List of roles retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create role
      description: Create a new role within a tenant
      tags:
        - Roles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...